Refactor to move the SentenceIdentificationService from within the SentenceCalculationService
