+++ src/test/kotlin/uk/gov/justice/digital/hmpps/calculatereleasedatesapi/service/SentenceCalculationServiceCalculateTest.kt
@@ -11,13 +11,14 @@ import java.time.LocalDate
 class SentenceCalculationServiceCalculateTest {
 
   private val sentenceCalculationService: SentenceCalculationService = SentenceCalculationService()
+  private val sentenceIdentificationService: SentenceIdentificationService = SentenceIdentificationService()
   private val jsonTransformation = JsonTransformation()
   private val offender = jsonTransformation.loadOffender("john_doe")
 
   @Test
   fun `Example 9`() {
     val sentence = jsonTransformation.loadSentence("2_year_sep_2013")
-    sentenceCalculationService.identify(sentence, offender)
+    sentenceIdentificationService.identify(sentence, offender)
     val offender = Offender("A1234BC", "John Doe", LocalDate.of(1980, 1, 1))
     val booking = Booking(offender, mutableListOf(sentence), mutableMapOf())
     val calculation = sentenceCalculationService.calculate(sentence, booking)
@@ -29,7 +30,7 @@ class SentenceCalculationServiceCalculateTest {
   @Test
   fun `Example 10`() {
     val sentence = jsonTransformation.loadSentence("3_year_dec_2012")
-    sentenceCalculationService.identify(sentence, offender)
+    sentenceIdentificationService.identify(sentence, offender)
     val offender = Offender("A1234BC", "John Doe", LocalDate.of(1980, 1, 1))
     val adjustments = mutableMapOf<AdjustmentType, Int>()
     adjustments[AdjustmentType.REMAND] = 35
@@ -45,7 +46,7 @@ class SentenceCalculationServiceCalculateTest {
   @Test
   fun `Example 11`() {
     val sentence = jsonTransformation.loadSentence("8_year_dec_2012")
-    sentenceCalculationService.identify(sentence, offender)
+    sentenceIdentificationService.identify(sentence, offender)
     val adjustments = mutableMapOf<AdjustmentType, Int>()
     adjustments[AdjustmentType.REMAND] = 10
     val offender = Offender("A1234BC", "John Doe", LocalDate.of(1980, 1, 1))
@@ -59,7 +60,7 @@ class SentenceCalculationServiceCalculateTest {
   @Test
   fun `Example 12`() {
     val sentence = jsonTransformation.loadSentence("8_year_feb_2015")
-    sentenceCalculationService.identify(sentence, offender)
+    sentenceIdentificationService.identify(sentence, offender)
     val adjustments = mutableMapOf<AdjustmentType, Int>()
     adjustments[AdjustmentType.REMAND] = 21
     val offender = Offender("A1234BC", "John Doe", LocalDate.of(1980, 1, 1))
